// client/app/ss-elements/common/ae-informationbar-itemtype.enum.ts
export enum AeInformationBarItemType {
    HolidaysAvailable = 1,
    TeamHolidays = 2,
    DocumentsAwaiting = 3,
    TasksToComplete = 4,
    TrainingCourses = 5,
    HolidayCountdown = 6,
    Joiners = 7,
    Documents = 8,
    TasksdueToday = 9,
    OutOfOffice = 10,
    NewTasks = 11,
    OverdueTasks = 12,
    IncompleteTasks = 13,
    DueTodayTask = 14,
    DueThisWeekTasks = 15,
    DueNextWeekTasks = 16,
    TrainingCertificates = 17,
    PersonalDocuments = 18,
    CompanyDocuments = 19
}